pyproject.toml
[project]
name = "intelligent_task_planner"
version = "0.1.0"
description = "An intelligent task scheduling application for students"
authors = [
    {name = "Rayed Riasat Rabbi", email = "rayed.rabbi@northsouth.edu"},
]
dependencies = [
    "django==5.2.4",
    "mysqlclient==2.2.7",
    "django-allauth[socialaccount]==65.10.0",
    "django-q2==1.8.0",
    "httpx==0.28.1",
    "pytest==8.4.1",
    "pytest-django==4.11.1",
    "django-environ==0.11.2",
    "ruff==0.8.1",
    "google-api-python-client>=2.179.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "pytz>=2025.2",
    "pyjwt>=2.10.1",
    "cryptography>=45.0.6",
]
requires-python = ">=3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["intelligent_task_planner"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "intelligent_task_planner.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--tb=short"
